requirements.txt
Flask
requests
groq
beautifulsoup4
python-dotenv